refactor remainder logic

+++ main.py
@@ -7,11 +7,12 @@ from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import ReplyKeyboardRemove
 
-from keyboards import delete_task_keyboard, hour_keyboard, minutes_keyboard, type_keyboard, week_day_keyboard, month_day_keyboard, \
+from src.keyboards import delete_task_keyboard, hour_keyboard, minutes_keyboard, type_keyboard, week_day_keyboard, month_day_keyboard, \
     confirm_keyboard
-from commands import bot_commands
-from config import API_TOKEN, remainder_types, db, timezone, scheduler
-from scheduler import add_tasks_from_db, add_task, delete_task, get_formatted_jobs, rm_all_tasks_from_db
+from src.commands import bot_commands
+from src.config import API_TOKEN, remainder_types, db, scheduler
+from src.scheduler import add_tasks_from_db, add_task, delete_task, get_formatted_jobs, rm_all_tasks_from_db, \
+    get_formatted_task
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
@@ -84,16 +85,11 @@ async def process_name(message: types.Message, state: FSMContext):
 async def process_name(message: types.Message, state: FSMContext):
     await state.update_data(text=message.text)
     data = await state.get_data()
-    await message.answer(
-        f"Вы добавили напомиание:\n"
-        f"Тип: {data['type']}\n"
-        f"День месяца: {data['month_day']}\n"
-        f"День недели: {data['week_day']}\n"
-        f"Время: {data['hour']}:{data['minutes']}\n"
-        f"Текст напоминания: {data['text']}",
-        reply_markup=confirm_keyboard(),
-    )
-    await state.set_state(Remainder.confirm)
+    data["user_id"] = message.from_user.id
+    data["task_id"] = add_task(data, bot)
+    db.insert(data)
+    await message.answer("Уведомление добавлено", reply_markup=ReplyKeyboardRemove())
+    await state.clear()
 
 @dp.message(Remainder.confirm)
 async def process_confirm(message: types.Message, state: FSMContext):
@@ -107,8 +103,6 @@ async def process_confirm(message: types.Message, state: FSMContext):
 
 @dp.message(Command("reminders"))
 async def cmd_start(message: types.Message, state: FSMContext):
-    for job in db.all():
-        print(job)
     formated_reminders = get_formatted_jobs()
     await message.answer(f"текущие напоминания:\n\n{formated_reminders}")
 
@@ -120,14 +114,31 @@ async def cmd_start(message: types.Message, state: FSMContext):
     
 @dp.message(Command("delete"))
 async def cmd_start(message: types.Message, state: FSMContext):
-    await message.answer(f"Какое напоминание удалить?", reply_markup=delete_task_keyboard(db.all()))
+    await message.answer(f"Какое напоминание удалить?", reply_markup=delete_task_keyboard())
     
     
-@dp.callback_query(lambda c: c.data.startswith("task_"))
+@dp.callback_query(lambda c: c.data.startswith("task__"))
 async def handle_task_selection(callback: types.CallbackQuery):
-    task_id = callback.data.split("_")[1]
-    delete_task(task_id)
-    await bot.send_message(callback.from_user.id, f"Напоминание {task_id} удалено")
+    task_id = callback.data.split("__")[1]
+    if task_id:
+        task = get_formatted_task(task_id)
+        await bot.send_message(
+            callback.from_user.id,
+            text=f"Хотите удалить уведомление\n{task}",
+            reply_markup=confirm_keyboard("task_delete", task_id),
+        )
+    else:
+        await bot.send_message(callback.from_user.id, "Отмена", reply_markup=ReplyKeyboardRemove())
+
+@dp.callback_query(lambda c: c.data.startswith("task_delete__"))
+async def handle_task_selection(callback: types.CallbackQuery):
+    should_delete = callback.data.split("__")[1]
+    if should_delete == "Да":
+        task_id = callback.data.split("__")[2]
+        delete_task(task_id)
+        await bot.send_message(callback.from_user.id, f"Напоминание удалено", reply_markup=ReplyKeyboardRemove())
+    else:
+        await bot.send_message(callback.from_user.id, "Отмена", reply_markup=ReplyKeyboardRemove())
 
 async def main():
     add_tasks_from_db(bot)
